move log dumping into platform actions
