+++ shared/actions/config.js
@@ -25,8 +25,6 @@ import {loggedInSelector} from '../constants/selectors'
 import {type AsyncAction} from '../constants/types/flux'
 import {type TypedState} from '../constants/reducer'
 import shallowEqual from 'shallowequal'
-import {quit} from '../util/quit-helper'
-import dumpLogs from '../logger/dump-log-fs'
 
 const maxAvatarsPerLoad = 50
 // TODO convert to sagas
@@ -335,15 +333,6 @@ function _onMobileAppStateChanged(action: ConfigGen.MobileAppStatePayload) {
   return Saga.put(ConfigGen.createChangedFocus({appFocused}))
 }
 
-function _dumpLogs(action: ConfigGen.DumpLogsPayload) {
-  dumpLogs().then(() => {
-    // quit as soon as possible
-    if (action.payload.reason === 'quitting through menu') {
-      quit('quitButton')
-    }
-  })
-}
-
 function* configSaga(): Saga.SagaGenerator<any, any> {
   yield Saga.safeTakeEveryPure(ConfigGen.bootstrapSuccess, _bootstrapSuccess)
   yield Saga.safeTakeEveryPure(ConfigGen.bootstrap, _bootstrap)
@@ -356,7 +345,6 @@ function* configSaga(): Saga.SagaGenerator<any, any> {
   yield Saga.safeTakeEveryPure(Chat2Gen.selectConversation, _setStartedDueToPush)
   yield Saga.safeTakeEveryPurePromise(ConfigGen.loadConfig, getConfig)
   yield Saga.safeTakeEveryPure(ConfigGen.mobileAppState, _onMobileAppStateChanged)
-  yield Saga.safeTakeEveryPure(ConfigGen.dumpLogs, _dumpLogs)
   yield Saga.fork(PlatformSpecific.platformConfigSaga)
 }
