--- shared/logger/dump-log-fs.js.flow
@@ -1,2 +1,0 @@
-// @flow
-declare export default function dumpLogs(): Promise<void>
